requirements.txt
fastapi==0.111.0
uvicorn[standard]>=0.31.1
websockets==12.0
onnxruntime>=1.24.0
numpy>=2.0.0
scipy>=1.14.0
librosa==0.10.2
soundfile==0.12.1
python-multipart==0.0.9
PyJWT==2.8.0
python-multipart==0.0.9
